Added interquartile range; renamed quartiles

+++ quartiles.py
@@ -5,7 +5,7 @@ def get_median(arr):
     if n%2==0:
         median=(arr[(n/2)-1]+arr[n/2])*0.5
     else:
-        median=(arr[n/2])
+        float(median=(arr[n/2]))
     return median
